% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\name{convertBase}
\alias{convertBase}
\title{Convert string representations of numbers in any base to any other base.}
\usage{
convertBase(x, base1 = 10, base2 = 10)
}
\arguments{
\item{x}{a vector of integers or strings to be converted}

\item{base1}{the base of x}

\item{base2}{the base of the output}
}
\description{
Convert string representations of numbers in any base to any other base.
}
\examples{

identical(convertBase(1234,base2=8),as.character(as.octmode(1234)))

convertBase(17771,base1=8,base2=30)
convertBase(17771,base1=8,base2=10)
convertBase(8185,base1=10,base2=30)

}
\seealso{
\code{\link{as.octmode}}

\code{\link{as.hexmode}}

\code{\link{strtoi}}
}

