% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grr.R
\docType{class}
\name{Grr-class}
\alias{Grr-class}
\alias{Grr}
\title{A Reference Class to represent a new data type for dataframe with grouped columns}
\description{
A Reference Class to represent a new data type for dataframe with grouped columns
}
\section{Fields}{

\describe{
\item{\code{groups}}{A list of a nemed vector of string (eg: list(grp1 = c("Field1", "Field3")))}

\item{\code{data}}{A data.frame object containing the data}
}}


\author{
Yannick Pascaud
}
