% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\docType{package}
\name{grr}
\alias{grr}
\alias{grr-package}
\title{Alternative Implementations of Base R Functions}
\description{
Alternative implementations of some base R functions, including sort, order, and match.  Functions are
simplified but can be faster or have other advantages.  See the documentation of individual functions
for details and benchmarks.
}
\details{
Note that these functions cannot be considered drop-in replacements for the functions in base \code{R}.  
They do not implement all the same parameters and do not work for all data types.  Utilize these 
with caution in specialized applications that require them.
}

