% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr.R
\docType{package}
\name{grr}
\alias{grr}
\alias{grr-package}
\title{Alternate Implementations of Base R Functions}
\description{
Alternate implementations of some base R functions, including sort, order, and match.  Functions are
faster and/or have been otherwise augmented.  See the documentation of individual functions
for details and benchmarks.
}

